Refactor server to send locale in body
