+++ package.json
@@ -43,6 +43,7 @@
     "favicons-webpack-plugin": "2.1.0",
     "html-webpack-plugin": "^3.2.0",
     "jest": "^24.9.0",
+    "jest-fetch-mock": "^3.0.1",
     "jest-styled-components": "^7.0.0",
     "node-sass": "^4.13.1",
     "react-test-renderer": "^16.12.0",
@@ -58,7 +59,6 @@
     "cookie-parser": "^1.4.4",
     "express": "^4.17.1",
     "immutable": "^4.0.0-rc.12",
-    "js-cookie": "^2.2.1",
     "prop-types": "^15.7.2",
     "react": "^16.12.0",
     "react-autosize-textarea": "^7.0.0",
