+++ server.js
@@ -23,11 +23,11 @@ app.listen(port, () => console.log(`Listening on port ${port}`));
 app.get('/api/locale', (req, res) => {
   const locale = detectLocale(req);
   console.log('/api/locale sending locale', locale);
-  res.cookie('locale', locale, { maxAge: (new Date() * 0.001) + (365 * 24 * 3600) });
-  res.status(200).end();
+  res.status(200)
+    .json({locale});
 });
 
 app.get('*', (req, res) =>{
-  res.sendFile(path.join(__dirname, '/client/build/index.html'));
-  res.status(200).end();
+  res.status(200)
+    .sendFile(path.join(__dirname, '/client/build/index.html'));
 });
