+++ test/setupJest.js
@@ -1,0 +1,7 @@
+// const fetchMock = require('jest-fetch-mock');
+
+// module.exports = () => {
+//   fetchMock.enableMocks();
+// };
+
+global.fetch = require('jest-fetch-mock');
